fix(user): 更新用户ID类型为uint64并调整相关代码

+++ internal/domain/base.go
@@ -3,7 +3,7 @@ package domain
 import "time"
 
 type Base struct {
-	ID        int
+	ID        uint64
 	CreatedAt *time.Time
 	UpdatedAt *time.Time
 	DeletedAt *time.Time
